some solution changes and add tests
